Small refactor and added support for standalone home-manager (#2)
* Renamed the configuration files in the hosts to default.nix

* Started refactoring flake.nix

Now I have a flakes utils that will build the files I need. It ow supports some standalone home-manager configurations, even though they are not done.

* Renamed user variable to username

* Moved all nixos-configs to utils

- All configs now use the mkNixosConfig utility
- Renamed `acer` to `nixos-acer` to match the hostname. It will stay like this while I think of a better name

* Refactored more functions and configs

Also updated some settings to work with standalone home-manager

* Initial setup to actually use wsl-work

Now I have the default shell environment working, and it possibly doesn't affect the other configurations

* Renamed a few folders

- `home` is now called `home-manager`
- `modules` is now called `nixos`

* Added a packages used by the work computer in wsl

* Removed arguments from modules where no arguement was used

* Added alpha as a start screen for neovim

* Added integration for the system clipboard

* Refactored some more things

Moved a neovim config to a file where it made more sense and changed the terminal font size

* Attempt to make nix LSP work properly

* Renamed a file

* Started implementing debugging

* Added more configs to wsl-work

Now I can run home-manager from the CLI, keychain will be started with the system, and added the aws cli tool

* Updates

- Updated the structure of flake.nix
- Updated the README.md file
- Small update to the nix.nix language config for neovim

* Added debugging for Flutter

- Configured the flutter plugin to use nvim_dap for debugging
- Tried to configure dap to debug rust, with little success

* Updated the README

* Added private nix configs submodule

Added a git submodule for nix configs so I can put configurations that I don't want to be public, such as information about my current work, which I don't know if can be made public or not. This way everything is in a different repository, which is a private repository in github

* Configure nix from home-manager

Home-manager will now configure nix and install it's own version

* Updated shortcuts

- Updated LSP find references to use Telescope instead
- Changed the leader key to space
- Changed the find file shortcut to find git files and the original find files was remaped
- Added several shortcuts to split the screen in the ways I like

* Updated flake.lock

* Updated README

Updated the README with instructions to install standalone home-manager
and small list of the profiles